# services/core/public/docs/openapi.yaml
openapi: 3.0.3

info:
  title: WeatherFlow Core API
  description: |
    API REST para la plataforma de servicios meteorológicos WeatherFlow.
    Este servicio gestiona usuarios, sus suscripciones y estaciones meteorológicas.
  version: 1.0.0

servers:
  - url:  http://127.0.0.1:8080/api
    description: Mismo host y puerto desde el que se abre Swagger UI (p. ej. http://localhost:8081)

tags:
  - name: Users
    description: Gestión de usuarios y sus suscripciones a estaciones
  - name: Weather Stations
    description: Gestión de estaciones meteorológicas

# =============================================================================
# PATHS
# =============================================================================

paths:

  # ---------------------------------------------------------------------------
  # USERS
  # ---------------------------------------------------------------------------

  /users:
    get:
      tags: [Users]
      summary: Listar todos los usuarios
      operationId: listUsers
      responses:
        '200':
          description: Lista de usuarios con sus suscripciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserWithSubscriptions'
              example:
                - id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "john@example.com"
                  firstName: "John"
                  lastName: "Doe"
                  subscriptions: []

    post:
      tags: [Users]
      summary: Crear un usuario
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              email: "john@example.com"
              first_name: "John"
              last_name: "Doe"
      responses:
        '201':
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithSubscriptions'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'

    get:
      tags: [Users]
      summary: Obtener un usuario por ID
      operationId: getUser
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithSubscriptions'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Users]
      summary: Actualizar un usuario
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            example:
              email: "john.updated@example.com"
              first_name: "John"
              last_name: "Updated"
      responses:
        '200':
          description: Usuario actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Users]
      summary: Eliminar un usuario
      operationId: deleteUser
      responses:
        '204':
          description: Usuario eliminado correctamente
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: El usuario tiene estaciones meteorológicas a su cargo y no puede eliminarse
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "User 550e8400-e29b-41d4-a716-446655440000 owns one or more weather stations and cannot be deleted."

  /users/{userId}/subscriptions:
    parameters:
      - name: userId
        in: path
        required: true
        description: UUID del usuario
        schema:
          type: string
          format: uuid

    post:
      tags: [Users]
      summary: Suscribir un usuario a una estación
      operationId: subscribeUserToStation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscribeRequest'
            example:
              station_id: "660e8400-e29b-41d4-a716-446655440001"
      responses:
        '200':
          description: Suscripción creada. Devuelve el usuario con la lista actualizada de suscripciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithSubscriptions'
        '404':
          description: Usuario o estación no encontrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '409':
          description: El usuario ya estaba suscripto a esa estación
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /users/{userId}/subscriptions/{stationId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: UUID del usuario
        schema:
          type: string
          format: uuid
      - name: stationId
        in: path
        required: true
        description: UUID de la estación
        schema:
          type: string
          format: uuid

    delete:
      tags: [Users]
      summary: Desuscribir un usuario de una estación
      description: Operación idempotente — devuelve 204 aunque el usuario no estuviera suscripto.
      operationId: unsubscribeUserFromStation
      responses:
        '204':
          description: Desuscripción exitosa (o el usuario ya no estaba suscripto)
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # WEATHER STATIONS
  # ---------------------------------------------------------------------------

  /stations:
    get:
      tags: [Weather Stations]
      summary: Listar estaciones meteorológicas
      description: |
        Devuelve todas las estaciones. Los filtros son opcionales y se pueden combinar.
        Sin parámetros, devuelve el listado completo.
      operationId: listStations
      parameters:
        - name: name
          in: query
          required: false
          description: Filtra por nombre parcial (búsqueda insensible a mayúsculas/minúsculas).
          schema:
            type: string
            maxLength: 255
          example: Central
        - name: created_from
          in: query
          required: false
          description: Incluye estaciones creadas en o después de esta fecha (inclusive).
          schema:
            type: string
            format: date
          example: "2026-02-01"
        - name: created_to
          in: query
          required: false
          description: Incluye estaciones creadas en o antes de esta fecha (inclusive). Debe ser mayor o igual que `created_from` si ambos se envían.
          schema:
            type: string
            format: date
          example: "2026-03-01"
      responses:
        '200':
          description: Lista de estaciones meteorológicas que coinciden con los filtros aplicados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Station'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    post:
      tags: [Weather Stations]
      summary: Crear una estación meteorológica
      operationId: createStation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStationRequest'
            example:
              owner_id: "550e8400-e29b-41d4-a716-446655440000"
              station_name: "Estación Central"
              latitude: -34.6037
              longitude: -58.3816
              sensor_model: "Davis Vantage Pro2"
              status: "active"
              climate_provider: "openweather"
      responses:
        '201':
          description: Estación creada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '422':
          description: Datos inválidos o el owner_id no corresponde a un usuario existente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /stations/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'

    get:
      tags: [Weather Stations]
      summary: Obtener una estación por ID
      operationId: getStation
      responses:
        '200':
          description: Estación encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Weather Stations]
      summary: Actualizar una estación meteorológica
      operationId: updateStation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStationRequest'
      responses:
        '200':
          description: Estación actualizada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Datos inválidos o el owner_id no corresponde a un usuario existente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

    delete:
      tags: [Weather Stations]
      summary: Eliminar una estación
      operationId: deleteStation
      responses:
        '204':
          description: Estación eliminada correctamente
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: La estación tiene mediciones asociadas y no puede eliminarse
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Station 660e8400-e29b-41d4-a716-446655440001 has measurements and cannot be deleted."

# =============================================================================
# COMPONENTS
# =============================================================================

components:

  # ---------------------------------------------------------------------------
  # PARAMETERS
  # ---------------------------------------------------------------------------

  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      description: UUID del recurso
      schema:
        type: string
        format: uuid

  # ---------------------------------------------------------------------------
  # RESPONSES
  # ---------------------------------------------------------------------------

  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
          example:
            message: "Resource not found."

    Conflict:
      description: Conflicto — el recurso ya existe con esos datos
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
          example:
            message: "The email 'john@example.com' is already in use."

    UnprocessableEntity:
      description: Error de validación
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
            example:
              message: "The email field is required."
              errors:
                email: ["The email field is required."]

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------

  schemas:

    # --- Users ----------------------------------------------------------------

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "john@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"

    UserWithSubscriptions:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            subscriptions:
              type: array
              items:
                $ref: '#/components/schemas/SubscriptionDetail'
              example: []

    SubscriptionDetail:
      type: object
      description: Detalle de una estación a la que el usuario está suscripto
      properties:
        stationId:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        name:
          type: string
          example: "Estación Central"
        latitude:
          type: number
          format: float
          example: -34.6037
        longitude:
          type: number
          format: float
          example: -58.3816
        sensorModel:
          type: string
          example: "Davis Vantage Pro2"
        status:
          type: string
          enum: [active, inactive]
          example: "active"
        climateProvider:
          type: string
          enum: [openweather]
          example: "openweather"

    CreateUserRequest:
      type: object
      required: [email, first_name, last_name]
      properties:
        email:
          type: string
          format: email
          example: "john@example.com"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"

    UpdateUserRequest:
      type: object
      required: [email, first_name, last_name]
      properties:
        email:
          type: string
          format: email
          example: "john.updated@example.com"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Updated"

    SubscribeRequest:
      type: object
      required: [station_id]
      properties:
        station_id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"

    # --- Weather Stations -----------------------------------------------------

    Station:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        ownerId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        stationName:
          type: string
          example: "Estación Central"
        latitude:
          type: number
          format: float
          minimum: -90
          maximum: 90
          example: -34.6037
        longitude:
          type: number
          format: float
          minimum: -180
          maximum: 180
          example: -58.3816
        sensorModel:
          type: string
          example: "Davis Vantage Pro2"
        status:
          type: string
          enum: [active, inactive]
          example: "active"
        climateProvider:
          type: string
          enum: [openweather]
          example: "openweather"

    CreateStationRequest:
      type: object
      required: [owner_id, station_name, latitude, longitude, sensor_model]
      properties:
        owner_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        station_name:
          type: string
          maxLength: 255
          example: "Estación Central"
        latitude:
          type: number
          format: float
          minimum: -90
          maximum: 90
          example: -34.6037
        longitude:
          type: number
          format: float
          minimum: -180
          maximum: 180
          example: -58.3816
        sensor_model:
          type: string
          maxLength: 255
          example: "Davis Vantage Pro2"
        status:
          type: string
          enum: [active, inactive]
          default: active
          example: "active"
        climate_provider:
          type: string
          enum: [openweather]
          default: openweather
          example: "openweather"

    UpdateStationRequest:
      type: object
      required: [owner_id, station_name, latitude, longitude, sensor_model, status]
      properties:
        owner_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        station_name:
          type: string
          maxLength: 255
          example: "Estación Central"
        latitude:
          type: number
          format: float
          minimum: -90
          maximum: 90
          example: -34.6037
        longitude:
          type: number
          format: float
          minimum: -180
          maximum: 180
          example: -58.3816
        sensor_model:
          type: string
          maxLength: 255
          example: "Davis Vantage Pro2"
        status:
          type: string
          enum: [active, inactive]
          example: "active"
        climate_provider:
          type: string
          enum: [openweather]
          default: openweather
          example: "openweather"